Move `ProfileClassJob` to the `classjob` package
